# backend/course-service/api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Course Management API",
        "description": "API for managing university courses and sections",
        "version": "1.0.0"
    },
    "paths": {
        "/courses/": {
            "post": {
                "tags": [
                    "courses"
                ],
                "summary": "Create Course",
                "description": "Create a new course\n\n- **course_id**: Unique identifier for the course\n- **title**: Title of the course\n- **description**: Description of the course\n- **instructor**: Course instructor name\n- **credits**: Number of credits\n- **department**: Department offering the course\n- **is_core**: Whether it's a core course\n- **level**: Course level (100, 200, etc.)\n- **prerequisites**: List of prerequisite course IDs\n- **categories**: List of category names\n- **sections**: List of sections for this course",
                "operationId": "create_course_courses__post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CourseCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Course"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response 404 Create Course Courses  Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "409": {
                        "description": "Course already exists",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": "error",
                                    "code": 409,
                                    "message": "Course already exists",
                                    "details": {
                                        "course_id": "CS101"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response 500 Create Course Courses  Post"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "courses"
                ],
                "summary": "Read Courses",
                "description": "Get all courses with pagination\n\n- **skip**: Number of records to skip\n- **limit**: Maximum number of records to return",
                "operationId": "read_courses_courses__get",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Course"
                                    },
                                    "title": "Response Read Courses Courses  Get"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response 404 Read Courses Courses  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response 422 Read Courses Courses  Get"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response 409 Read Courses Courses  Get"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response 500 Read Courses Courses  Get"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/courses/{course_id}": {
            "get": {
                "tags": [
                    "courses"
                ],
                "summary": "Read Course",
                "description": "Get a specific course by ID\n\n- **course_id**: ID of the course to retrieve",
                "operationId": "read_course_courses__course_id__get",
                "parameters": [
                    {
                        "name": "course_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Course Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Course"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course not found",
                        "content": {
                            "application/json": {
                                "example": {
                                    "status": "error",
                                    "code": 404,
                                    "message": "Course not found",
                                    "details": {
                                        "course_id": "UNKNOWN101"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response 422 Read Course Courses  Course Id  Get"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response 409 Read Course Courses  Course Id  Get"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response 500 Read Course Courses  Course Id  Get"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "courses"
                ],
                "summary": "Update Course",
                "description": "Update a course\n\n- **course_id**: ID of the course to update\n- **course**: Updated course data",
                "operationId": "update_course_courses__course_id__put",
                "parameters": [
                    {
                        "name": "course_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Course Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CourseCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Course"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response 409 Update Course Courses  Course Id  Put"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response 500 Update Course Courses  Course Id  Put"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "courses"
                ],
                "summary": "Delete Course",
                "description": "Delete a course\n\n- **course_id**: ID of the course to delete",
                "operationId": "delete_course_courses__course_id__delete",
                "parameters": [
                    {
                        "name": "course_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Course Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "404": {
                        "description": "Course not found"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response 422 Delete Course Courses  Course Id  Delete"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response 409 Delete Course Courses  Course Id  Delete"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "title": "Response 500 Delete Course Courses  Course Id  Delete"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Course": {
                "properties": {
                    "course_id": {
                        "type": "string",
                        "title": "Course Id"
                    },
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "instructor": {
                        "type": "string",
                        "title": "Instructor"
                    },
                    "credits": {
                        "type": "integer",
                        "title": "Credits"
                    },
                    "department": {
                        "type": "string",
                        "title": "Department"
                    },
                    "is_core": {
                        "type": "boolean",
                        "title": "Is Core",
                        "default": false
                    },
                    "level": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Level"
                    },
                    "prerequisites": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Prerequisites",
                        "default": []
                    },
                    "sections": {
                        "items": {
                            "$ref": "#/components/schemas/Section"
                        },
                        "type": "array",
                        "title": "Sections",
                        "default": []
                    },
                    "categories": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Categories",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "course_id",
                    "title",
                    "description",
                    "instructor",
                    "credits",
                    "department"
                ],
                "title": "Course"
            },
            "CourseCreate": {
                "properties": {
                    "course_id": {
                        "type": "string",
                        "title": "Course Id"
                    },
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description"
                    },
                    "instructor": {
                        "type": "string",
                        "title": "Instructor"
                    },
                    "credits": {
                        "type": "integer",
                        "title": "Credits"
                    },
                    "department": {
                        "type": "string",
                        "title": "Department"
                    },
                    "is_core": {
                        "type": "boolean",
                        "title": "Is Core",
                        "default": false
                    },
                    "level": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Level"
                    },
                    "prerequisites": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Prerequisites",
                        "default": []
                    },
                    "sections": {
                        "items": {
                            "$ref": "#/components/schemas/SectionCreate"
                        },
                        "type": "array",
                        "title": "Sections",
                        "default": []
                    },
                    "categories": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Categories",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "course_id",
                    "title",
                    "description",
                    "instructor",
                    "credits",
                    "department"
                ],
                "title": "CourseCreate"
            },
            "Section": {
                "properties": {
                    "section_id": {
                        "type": "string",
                        "title": "Section Id"
                    },
                    "instructor": {
                        "type": "string",
                        "title": "Instructor"
                    },
                    "capacity": {
                        "type": "integer",
                        "title": "Capacity"
                    },
                    "schedule_day": {
                        "type": "string",
                        "title": "Schedule Day"
                    },
                    "schedule_time": {
                        "type": "string",
                        "title": "Schedule Time"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "course_id": {
                        "type": "string",
                        "title": "Course Id"
                    }
                },
                "type": "object",
                "required": [
                    "section_id",
                    "instructor",
                    "capacity",
                    "schedule_day",
                    "schedule_time",
                    "id",
                    "course_id"
                ],
                "title": "Section"
            },
            "SectionCreate": {
                "properties": {
                    "section_id": {
                        "type": "string",
                        "title": "Section Id"
                    },
                    "instructor": {
                        "type": "string",
                        "title": "Instructor"
                    },
                    "capacity": {
                        "type": "integer",
                        "title": "Capacity"
                    },
                    "schedule_day": {
                        "type": "string",
                        "title": "Schedule Day"
                    },
                    "schedule_time": {
                        "type": "string",
                        "title": "Schedule Time"
                    }
                },
                "type": "object",
                "required": [
                    "section_id",
                    "instructor",
                    "capacity",
                    "schedule_day",
                    "schedule_time"
                ],
                "title": "SectionCreate"
            }
        }
    }
}